chore(frontends/lean/info_manager): make copyable and integrate into snapshots
