+++ src/frontends/lean/CMakeLists.txt
@@ -7,4 +7,4 @@ pp.cpp structure_cmd.cpp structure_instance.cpp
 init_module.cpp type_util.cpp local_ref_info.cpp decl_attributes.cpp
 opt_cmd.cpp prenum.cpp print_cmd.cpp elaborator.cpp elaborator_exception.cpp
 match_expr.cpp local_context_adapter.cpp decl_util.cpp definition_cmds.cpp
-brackets.cpp tactic_notation.cpp)
+brackets.cpp tactic_notation.cpp info_manager.cpp)
