+++ src/frontends/lean/elaborator.h
@@ -157,6 +157,7 @@ private:
     expr visit_sort(expr const & e);
     expr visit_const_core(expr const & e);
     expr ensure_function(expr const & e);
+    void save_identifier_info(expr const & f);
     expr visit_function(expr const & fn, bool has_args, expr const & ref);
     format mk_app_type_mismatch_error(expr const & t, expr const & arg, expr const & arg_type, expr const & expected_type);
     format mk_too_many_args_error(expr const & fn_type);
